Address review feedback

+++ python/cuda_parallel/cuda/parallel/experimental/_bindings.pyx
@@ -1630,7 +1630,7 @@ cdef extern from "cccl/c/transform.h":
       cccl_device_transform_build_result_t build,
       cccl_iterator_t d_in,
       cccl_iterator_t d_out,
-      unsigned long long num_items,
+      uint64_t num_items,
       cccl_op_t op,
       CUstream stream) nogil
 
@@ -1648,7 +1648,7 @@ cdef extern from "cccl/c/transform.h":
       cccl_iterator_t d_in1,
       cccl_iterator_t d_in2,
       cccl_iterator_t d_out,
-      unsigned long long num_items,
+      uint64_t num_items,
       cccl_op_t op,
       CUstream stream) nogil
